Refactor Campaign Service methods
- Add separate method to get campaign organisations
- Move similar methods from Campaign model to Campaign service
- Add new method to return CampaignListDTO
- Remove organisation detail from returned campaign
